Completata la parte relativa alla visualizzazione delle ore da usare per recuperare i 91CE

+++ app/models/dto/AbsenceToRecoverDto.java
@@ -11,6 +11,6 @@ public class AbsenceToRecoverDto {
   public LocalDate recoverDate;
   public int quantityRecovered;
   public int quantityToRecover;
-  public double percentage;
+  public float percentage;
 
 }
